Add support for deploying a `CiliumNetworkPolicy` for RabbitMQ
This commit extends the component to deploy a `CiliumNetworkPolicy`
which allows traffic from outside the cluster to reach the RabbitMQ
cluster over the LB service.

This policy is only deployed when Commodore component `cilium` is
detected for the cluster on which the component is deployed. Note that
the policy is deployed even when the RabbitMQ cluster isn't exposed on a
LB service.